Custom steps can disable their own caching
This is a design decision for steps creators, not end-users.

For some steps, like data-loading and exporting, step caching does not
makes sense as it duplicates the data without modifying it.

Now, individual steps can override their caching, without impact to the
re-computation mechanism: having no cache does NOT mean we need to
re-compute the subsequent steps because it does NOT means something was
changed. It's the user's responsibility to turn off caching if they want to.